# BackendV2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FUTCOL API",
        "description": "API para el sistema de gestión de torneos de fútbol FUTCOL. Esta API permite gestionar torneos, equipos, jugadores, amonestaciones y estadísticas.",
        "contact": {
            "name": "FUTCOL Support Team",
            "email": "admin@futcol.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "FUTCOL API Server"
        }
    ],
    "paths": {
        "/api/amonestaciones": {
            "get": {
                "tags": [
                    "Amonestaciones"
                ],
                "summary": "Obtener todas las amonestaciones",
                "operationId": "8a2b7f1c70fcf4c2ef0181dcb45c4a33",
                "responses": {
                    "200": {
                        "description": "Lista de amonestaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "Amonestaciones"
                ],
                "summary": "Crear una nueva amonestación",
                "operationId": "c4e75e8329595de75a755974af2c2866",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "jugador_id",
                                    "equipo_id",
                                    "encuentro_id",
                                    "numero_camiseta"
                                ],
                                "properties": {
                                    "jugador_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "equipo_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "encuentro_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "numero_camiseta": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "tarjeta_roja": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "tarjeta_amarilla": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "tarjeta_azul": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Amonestación creada correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/amonestaciones/{id}": {
            "get": {
                "tags": [
                    "Amonestaciones"
                ],
                "summary": "Obtener una amonestación específica",
                "operationId": "8bcee443bd6d315b72e8aa7a01f1320e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la amonestación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la amonestación"
                    },
                    "404": {
                        "description": "Amonestación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Amonestaciones"
                ],
                "summary": "Actualizar una amonestación",
                "operationId": "f15e80dd6604993950b3b7564ec9cc26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la amonestación a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jugador_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "equipo_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "encuentro_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "numero_camiseta": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "tarjeta_roja": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "tarjeta_amarilla": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "tarjeta_azul": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Amonestación actualizada correctamente"
                    },
                    "404": {
                        "description": "Amonestación no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Amonestaciones"
                ],
                "summary": "Eliminar una amonestación",
                "operationId": "1e38b711baf9e84c096e360818533338",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la amonestación a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Amonestación eliminada exitosamente"
                    },
                    "404": {
                        "description": "Amonestación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT en el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}